chore: Split dashboard test groups out into their own spec files
