+++ src/e2e/specs/dashboard/dashboard-content.spec.ts
@@ -1,0 +1,89 @@
+/* This Source Code Form is subject to the terms of the Mozilla Public
+ * License, v. 2.0. If a copy of the MPL was not distributed with this
+ * file, You can obtain one at http://mozilla.org/MPL/2.0/. */
+
+import { test, expect } from "../../fixtures/basePage.js";
+import { checkAuthState } from "../../utils/helpers.js";
+
+// bypass login
+test.use({ storageState: "./e2e/storageState.json" });
+
+test.describe(`${process.env.E2E_TEST_ENV} - Breaches Dashboard - Content @smoke`, () => {
+  test.beforeEach(async ({ dashboardPage, page }) => {
+    await dashboardPage.open();
+
+    try {
+      await checkAuthState(page);
+    } catch {
+      console.log("[E2E_LOG] - No fxa auth required, proceeding...");
+    }
+  });
+
+  test("Verify that the exposure list contains action needed", async ({
+    dashboardPage,
+    page,
+  }) => {
+    // link to testrail
+    test.info().annotations.push({
+      type: "testrail",
+      description:
+        "https://testrail.stage.mozaws.net/index.php?/cases/view/2301533",
+    });
+
+    await expect(dashboardPage.exposuresHeading).toBeVisible();
+    const listCount = await page
+      .locator('//div[starts-with(@class, "StatusPill_pill")]')
+      .count();
+
+    // verify exposure list conatins only exposures that need to be fixed
+    if (listCount > 0) {
+      for (let i = 0; i < listCount; i++) {
+        await expect(
+          page.locator(
+            `(//div[starts-with(@class, 'StatusPill_pill')])[${i + 1}]`,
+          ),
+        ).toHaveText("Action needed");
+      }
+    }
+  });
+
+  test("Verify that the exposure list contains only fixed and in progress cards", async ({
+    dashboardPage,
+    page,
+  }) => {
+    // link to testrail
+    test.info().annotations.push(
+      {
+        type: "testrail",
+        description:
+          "https://testrail.stage.mozaws.net/index.php?/cases/view/2301533",
+      },
+      {
+        type: "testrail",
+        description:
+          "https://testrail.stage.mozaws.net/index.php?/cases/view/2546463",
+      },
+    );
+
+    await expect(dashboardPage.exposuresHeading).toBeVisible();
+    await dashboardPage.fixedTab.click();
+
+    // verify fixed or in-progress
+    await expect(dashboardPage.fixedHeading).toBeVisible();
+
+    // TODO: add stub to fill in fixed/in-progress items
+    const listCount = await page
+      .locator('//div[starts-with(@class, "StatusPill_pill")]')
+      .count();
+    // verify exposure list contains only exposures that need to be fixed
+    if (listCount > 0) {
+      for (let i = 0; i < listCount; i++) {
+        await expect(
+          page.locator(
+            `(//div[starts-with(@class, "StatusPill_pill")])[${i + 1}]`,
+          ),
+        ).toHaveText(/In progress|Fixed/);
+      }
+    }
+  });
+});
